See description - also cleaned up the gitignore file.
What I have completed is:
When you load the home page the cameras populate the config table (provided the cameras aren't a;ready there. (I will need to have it return the current config if they are there i.e. status true or false.)
Master wipe using backend api end point /hardreset
Have created backend api to activate and deactivate the status (I need to enable the backend to send the first message to the queue next)

My todo list:
Enable the backend to send the first message to the SQS queue.
add a disable button and map it to the cancel end point.
Build a backend end point to get the counts as a roll up.
Build a page to see the data for each camera