too much to summarise!
So I think everything is working now,
I think that what we need to do next is have it up and working in AWS.

I introduced a LOT of bugs by bringing in the redis cache but I have overcome them.
bugs were:
updating the status meant that the status was not showing the right status (due to the page now using the cache
Had to workout how to iterate over the cache values and change the particular value (not simple when everything is one key and redis doesn't allow sub keys)
Had to do the same for delete.
Had a VERY unique issue where I truncated the database to play with the cache not realising I needed to flush the cache as it still had what was originally there and thus the database never became populated I spent A LOT longer than I care to admit on that!
because we can't simply hard reset the cache like I can locally I built into the function hard reset to auto clear the cache.

I fixed the styling to pop a bit more added a nav bar to give the appearance that I care about styling
In my honest opinion I have styling and designing it is the bane of my existence whenever I have to do it...
I added new end points for the data table stuff

todo:
Deploy to AWS and pray.

I am going to do that today after some sleep. I will deploy and take a bunch of screenshots. But I am at a point I am happy to show people this now 🙂